Move TLS configuration to a submodule.
